# src/lib/requests_bodies/UserUpdateRequestBody.ts
import { IsString, IsOptional, MinLength } from "class-validator";

export default class UserUpdateRequestBody {
  @IsOptional()
  @IsString()
  firstName?: string;

  @IsOptional()
  @IsString()
  lastName?: string;

  @IsOptional()
  @IsString()
  @MinLength(8)
  password?: string;
}